feat: 컬럼 추가로 인한 Shopping 수정

+++ feature/shopping-cart/src/main/java/com/project/shopping_cart/ShoppingCartViewModel.kt
@@ -130,7 +130,7 @@ class ShoppingCartViewModel @Inject constructor(
                 insertShoppingCartItemUseCase.invoke(
                     ShoppingCart(
                         name = state.addItemNameQuery,
-                        count = state.addItemCount.toInt(),
+                        count = state.addItemCount.toDouble(),
                         category = getIndexToIngredientCategory(state.addItemCategorySelected),
                         success = false
                     )
